Move several extension methods into OmniSharp.Utilities namespace
